requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
pydantic
geoalchemy2
alembic
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
